feat: 补充集成测试、文档更新
